refactor: three architecture cleanups from the deepening sweep
Consolidate scattered logic surfaced while reviewing the archive-color work:

1. Name the archive-indicator color once. The orange Color("208") was
   hand-copied into review.go (banner), list.go (dot), and schedule.go
   (timeline slug). Extract `archiveColor` in goaltable.go, referenced by all
   three. Kept independent of UrgencyDueToday's coincidentally-equal 208.

2. Route every credentialed command through loadClient. Seven commands
   hand-rolled the ConfigExists -> LoadConfig -> NewHTTPClient preamble; two
   of them (list, filtered) printed the "no config" error to stdout while the
   rest used stderr. Widen loadClient to (*Config, Client, bool) so view/review
   get the config they need, delete all seven inline copies, and fix the
   stdout/stderr divergence in one move. (loadConfigAndGoals left alone: it
   returns wrapped errors for next/schedule rather than printing.)

3. Extract the goal-detail formatting library into goaldetail.go. The 8 pure
   format* functions shared by view, review, and the TUI grid lived in
   review.go, so "how is a goal detail formatted?" was answered in a file named
   after only one of its callers. Pure move, no behavior change.

Build, go vet, and full test suite green.
