+++ add.go
@@ -35,7 +35,7 @@ func handleAddCommand() {
 		os.Exit(code)
 	}
 
-	client, ok := loadClient(os.Stderr)
+	_, client, ok := loadClient(os.Stderr)
 	if !ok {
 		os.Exit(1)
 	}
