+++ api.go
@@ -42,18 +42,10 @@ func (k *keyValueFlag) Set(v string) error {
 
 // handleAPICommand makes a raw, authenticated request to the Beeminder API.
 func handleAPICommand() {
-	if !ConfigExists() {
-		fmt.Fprintln(os.Stderr, "Error: No configuration found. Please run 'buzz auth login' to authenticate.")
+	_, client, ok := loadClient(os.Stderr)
+	if !ok {
 		os.Exit(1)
 	}
-
-	config, err := LoadConfig()
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error: Failed to load config: %s\n", redactError(err))
-		os.Exit(1)
-	}
-
-	client := NewHTTPClient(config)
 	os.Exit(runAPICommand(os.Args[2:], client, os.Stdout, os.Stderr))
 }
