+++ create.go
@@ -82,7 +82,7 @@ func handleCreateCommand() {
 		}
 	}
 
-	client, ok := loadClient(os.Stderr)
+	_, client, ok := loadClient(os.Stderr)
 	if !ok {
 		os.Exit(1)
 	}
