+++ data.go
@@ -14,7 +14,7 @@ import (
 
 // handleDataCommand lists a goal's datapoints.
 func handleDataCommand() {
-	client, ok := loadClient(os.Stderr)
+	_, client, ok := loadClient(os.Stderr)
 	if !ok {
 		os.Exit(1)
 	}
