+++ deadline_cmd.go
@@ -31,7 +31,7 @@ func handleDeadlineCommand() {
 		os.Exit(code)
 	}
 
-	client, ok := loadClient(os.Stderr)
+	_, client, ok := loadClient(os.Stderr)
 	if !ok {
 		os.Exit(1)
 	}
