+++ uncle.go
@@ -41,18 +41,10 @@ func handleUncleCommand() {
 	goalSlug := args[0]
 	skipConfirm := *yes || *yesShort
 
-	if !ConfigExists() {
-		fmt.Fprintln(os.Stderr, "Error: No configuration found. Please run 'buzz auth login' to authenticate.")
+	_, client, ok := loadClient(os.Stderr)
+	if !ok {
 		os.Exit(1)
 	}
-
-	config, err := LoadConfig()
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "Error: Failed to load config: %s\n", redactError(err))
-		os.Exit(1)
-	}
-
-	client := NewHTTPClient(config)
 
 	if !skipConfirm {
 		fmt.Printf("Call uncle on %s? This will instantly derail the goal and charge the pledge. [y/N] ", goalSlug)
